Create filter_scary_words.py
This regular expression will match any word that contains one of the words on the scary_words_list variable. The sub() function will then replace all the matches with an empty string, which will effectively remove the scary words from the text.

This is just one way to filter out words that would paint likenesses that people would find disturbing. There are many other ways to do this, and the best way will depend on the specific context.